Added persistent configuration of computer.
Configuration is now stored in the first page of flash memory.
This is read on startup.

The default values are defined in the configuration.h file.

most other files have been updated to use the values from the
configuration.
Still need to update the dataLogging transmit vs dave to memory to use
flag from config. Also change start address for memory and end address.

Need to still update xtract to be able to change these parameters.

- Changed headers to fix problems. Now each file includes the headers it
needs explicitly, main.h is not really used anymore.

-Added parameter sector erase function to flash module.
-added read/write configuration functions to the configuration module.